feat: gorm modelsの切り出し
